Revert "apiauth and sessionauth middleware created"
