+++ routes/user.js
@@ -3,7 +3,18 @@ const models = require('../models');
 const helper = require("../utils/helper");
 const env = process.env.NODE_ENV || 'development';
 const config = require("../config/config")[env];
-
+router.use(function(req, res, next) {
+  const api = req.get('X-API-KEY');
+  console.log(api);
+  if (api == config.api_key)
+    next()
+  else
+    res.send({
+      "error": true,
+      "status": "FAILURE",
+      "message": "Invalid API KEY"
+    });
+});
 router.get("/", function(req, res) {
   senduserdetails(req, (val) => {
     res.send(val);
